more debugging and testing on SQLite

+++ Classes/CSVGenerators/CSVGeneratorBase.cs
@@ -16,7 +16,7 @@ namespace CHaMPWorkbench.Classes.CSVGenerators
         }
 
         // All CSV generators must implement the run method that actually performs the CSV generation
-        public abstract System.IO.FileInfo Run(int nVisitID, string sFilePath);
+        public abstract System.IO.FileInfo Run(long nVisitID, string sFilePath);
 
         protected string AddNumericField(ref SQLiteDataReader dbRead, string sFieldName)
         {
